requirements.txt
fastapi
passlib
dotenv
python-jose==3.4.0
uvicorn
pydantic
python-multipart